pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghost"
version = "0.1.0"
description = "GHOST: sovereign adaptive knowledge system (retrieval + optimization + orchestration)"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "numpy>=1.26",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pydantic>=2.0",
    "typer>=0.12",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=3.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
ghost = "ghost_cli.main:app"

[tool.setuptools]
packages = [
    "ghost_core",
    "ghost_retrieval",
    "ghost_optimizer",
    "ghost_orchestrator",
    "ghost_orchestrator.adapters",
    "ghost_governance",
    "ghost_api",
    "ghost_cli",
]

[tool.pytest.ini_options]
testpaths = ["ghost_tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
